refactor(fmtutil): move quota types and format functions from usage
Move QuotaWindow, ExhaustedWindow, ExtraUsage, Data types and
FormatQuotaWindow, FormatStaleQuotaWindow, FormatRateLimitSegment,
FindExhaustedWindow functions into fmtutil. These are needed by both
the stdin path and the API path in the upcoming stdin-first architecture.
