+++ internal/fmtutil/format_test.go
@@ -3,6 +3,7 @@ package fmtutil
 import (
 	"strings"
 	"testing"
+	"time"
 )
 
 func TestDuration(t *testing.T) {
@@ -148,6 +149,167 @@ func TestJoinPipe(t *testing.T) {
 
 			if got := JoinPipe(tt.parts); got != tt.want {
 				t.Errorf("JoinPipe(%v) = %q, want %q", tt.parts, got, tt.want)
+			}
+		})
+	}
+}
+
+func TestFormatQuotaWindow(t *testing.T) {
+	t.Parallel()
+
+	win := &QuotaWindow{
+		Utilization:      45.3,
+		ResetsAt:         time.Now().Add(3 * time.Hour),
+		TotalMinutes:     10080,
+		RemainingMinutes: 6857,
+	}
+
+	got := FormatQuotaWindow(win, "7d", "")
+	if got == "" {
+		t.Error("FormatQuotaWindow returned empty string")
+	}
+
+	if !strings.Contains(got, "7d") || !strings.Contains(got, "45%") {
+		t.Errorf("FormatQuotaWindow = %q, missing expected content", got)
+	}
+}
+
+func TestFormatStaleQuotaWindow(t *testing.T) {
+	t.Parallel()
+
+	win := &QuotaWindow{
+		Utilization:      45.3,
+		ResetsAt:         time.Now().Add(3 * time.Hour),
+		TotalMinutes:     10080,
+		RemainingMinutes: 6857,
+	}
+
+	got := FormatStaleQuotaWindow(win, "7d", "")
+	if !strings.Contains(got, "?%") {
+		t.Errorf("FormatStaleQuotaWindow should contain '?%%', got %q", got)
+	}
+}
+
+func TestFindExhaustedWindow(t *testing.T) {
+	t.Parallel()
+
+	got := FindExhaustedWindow(&Data{
+		FiveHour: &QuotaWindow{
+			Utilization:      88,
+			RemainingMinutes: 45,
+		},
+		SevenDay: &QuotaWindow{
+			Utilization:      99,
+			RemainingMinutes: 125,
+		},
+	}, false)
+
+	if got == nil {
+		t.Fatal("expected non-nil ExhaustedWindow")
+	}
+
+	if got.Name != "7d" {
+		t.Errorf("Name = %q, want %q", got.Name, "7d")
+	}
+
+	if got.Minutes != 125 {
+		t.Errorf("Minutes = %d, want %d", got.Minutes, 125)
+	}
+}
+
+func TestFindExhaustedWindowFiveHour(t *testing.T) {
+	t.Parallel()
+
+	got := FindExhaustedWindow(&Data{
+		FiveHour: &QuotaWindow{
+			Utilization:      100,
+			RemainingMinutes: 30,
+		},
+		SevenDay: &QuotaWindow{
+			Utilization:      50,
+			RemainingMinutes: 125,
+		},
+	}, false)
+
+	if got == nil {
+		t.Fatal("expected non-nil ExhaustedWindow")
+	}
+
+	if got.Name != "5h" {
+		t.Errorf("Name = %q, want %q", got.Name, "5h")
+	}
+}
+
+func TestFindExhaustedWindowPerModel(t *testing.T) {
+	t.Parallel()
+
+	data := &Data{
+		SevenDay: &QuotaWindow{
+			Utilization:      50,
+			RemainingMinutes: 125,
+		},
+		SevenDayOpus: &QuotaWindow{
+			Utilization:      100,
+			RemainingMinutes: 60,
+		},
+	}
+
+	got := FindExhaustedWindow(data, true)
+	if got == nil {
+		t.Fatal("expected non-nil ExhaustedWindow")
+	}
+
+	if got.Name != "7d-opus" {
+		t.Errorf("Name = %q, want %q", got.Name, "7d-opus")
+	}
+
+	gotDisabled := FindExhaustedWindow(data, false)
+	if gotDisabled != nil {
+		t.Errorf("expected nil when perModel is false, got %+v", gotDisabled)
+	}
+}
+
+func TestFindExhaustedWindowNil(t *testing.T) {
+	t.Parallel()
+
+	if got := FindExhaustedWindow(nil, false); got != nil {
+		t.Error("expected nil for nil data")
+	}
+}
+
+func TestFindExhaustedWindowNoExhausted(t *testing.T) {
+	t.Parallel()
+
+	got := FindExhaustedWindow(&Data{
+		FiveHour: &QuotaWindow{Utilization: 50, RemainingMinutes: 45},
+		SevenDay: &QuotaWindow{Utilization: 60, RemainingMinutes: 125},
+	}, false)
+
+	if got != nil {
+		t.Error("expected nil when no window is exhausted")
+	}
+}
+
+func TestFormatRateLimitSegment(t *testing.T) {
+	t.Parallel()
+
+	tests := []struct {
+		name     string
+		input    *ExhaustedWindow
+		expected string
+	}{
+		{"nil window", nil, "⛔ limit hit"},
+		{"5h with time", &ExhaustedWindow{Name: "5h", Minutes: 134}, "⛔ 5h limit hit (2h 14m)"},
+		{"7d with time", &ExhaustedWindow{Name: "7d", Minutes: 1440}, "⛔ 7d limit hit (1d 0h)"},
+		{"zero minutes", &ExhaustedWindow{Name: "5h", Minutes: 0}, "⛔ 5h limit hit"},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			t.Parallel()
+
+			if got := FormatRateLimitSegment(tt.input); got != tt.expected {
+				t.Errorf("FormatRateLimitSegment() = %q, want %q", got, tt.expected)
 			}
 		})
 	}
